Fix emqttd mgmt;Refine process tree structure
- `Cog.BusDriver` now traps exits if it successfully starts
  `emqttd`. This allows Cog to keep the message bus state in sync with
  Cog's overall state. If Cog is up, the message bus is up. If Cog is
  down, the message bus is down.

- Removed the `mnesia` data file deletion logic from
  `Cog.BusDriver. Wound up causing more problems than it solved. We can
  tell users to delete the data dir in the release notes.

- Moved all of Cog core, modulo `Cog.Repo` and `Cog.BusDriver`, to a
  separate supervisor named `Cog.CoreSup` with a restart strategy of
  one-for-one. The top level supervisor is now responsible for
  `Cog.Repo`, `Cog.BusDriver`, and `Cog.CoreSup` and uses the
  one-for-all restart strategy. This should stabilize Cog as all
  processes will be restarted if the database or message bus goes down.

Fixes #897
